refactor: finish the snapshot-store file renames
# Summary

The two files that kept the crate-name prefix are renamed, and the
`OpenSnapshot` trait is reunited with its only implementation.

# Details

`snapshot_store_v003.rs` became `store.rs` when the crate was extracted,
but `snapshot_store_error.rs` and `snapshot_store_test.rs` kept spelling
the crate name inside the crate. They are now `error.rs` and
`store_test.rs`.

`db_open_snapshot_impl.rs` held nothing but `impl OpenSnapshot for DB`,
46 lines away from the 28-line trait it implements. Both now live in
`open_snapshot.rs`; the module stays public and its path is unchanged
for the two external callers.
