+++ crates/server/snapshot-store/src/error.rs
