+++ crates/server/snapshot-store/src/store_test.rs
